api/requirements.txt
fastapi
uvicorn
python-multipart
jinja2
markdown2
python-dotenv
google-generativeai
supabase
